Refatoração Update
- Adicionado detalhes.
- Melhorado documentação.
